Backend/Muffakir-V1-main/requirements.txt
transformers
langchain
langchain-community
unstructured[pdf]
chromadb
sentence-transformers
rank-bm25
nltk
openai
groq
langchain-groq
gradio
tiktoken
fastembed
langchain-chroma
streamlit
crewai
tavily-python
langchain_openai
streamlit_chat